requirements.txt
Flask==2.1.1
opencv-python==4.5.5.64
face_recognition==1.3.0
numpy==1.21.2
Pillow==8.4.0
Werkzeug==2.1.1
requests==2.26.0
sqlalchemy==1.4.25